Refactor component structure and imports across project directories
