+++ components/apps/ai-chat/index.ts
@@ -1,0 +1,2 @@
+export { default } from './ai-chat-console'
+
