+++ components/apps/auth/index.ts
@@ -1,0 +1,2 @@
+export { default } from './supabase-login-app'
+
