+++ components/apps/csv-parser/csv-parser-app.tsx
@@ -716,3 +716,5 @@ export function CsvParserApp() {
     </div>
   )
 }
+
+export default CsvParserApp
