+++ components/apps/csv-parser/index.ts
@@ -1,0 +1,2 @@
+export { default } from './csv-parser-app'
+
