+++ components/shared/about-this-desktop-app.tsx
