+++ components/shared/credits-manager.tsx
